Qiskit tests combined
Added qiskit tests to test_zne
